fix warn.

+++ frameworks/core/components_ng/pattern/text/text_pattern.cpp
@@ -3459,7 +3459,7 @@ bool TextPattern::DidExceedMaxLines() const
 TextLineMetrics TextPattern::GetLineMetrics(int32_t lineNumber)
 {
     CHECK_NULL_RETURN(pManager_, TextLineMetrics());
-    if (lineNumber < 0 || lineNumber > GetLineCount()) {
+    if (lineNumber < 0 || static_cast<size_t>(lineNumber) > GetLineCount()) {
         TAG_LOGI(AceLogTag::ACE_TEXT, "GetLineMetrics failed, lineNumber not between 0 and max lines:%{public}d",
             lineNumber);
         return TextLineMetrics();
